requirements.txt
pyrofork==2.3.69
tgcrypto-pyrofork
pymongo==4.13.2
motor==3.7.1
marshmallow<4.0.0
umongo
requests
bs4
shortzy
pytz
aiohttp
telegraph
pymediainfo-pyrofork

# Other tools
opencv-python-headless==4.8.1.78
python-dotenv==0.21.1
PyLeaves==1.0.2
humanize
Pillow
psutil
numpy
wheel
ujson
wget
aiofiles
dnspython==2.7.0
python-decouple
ffmpeg-python==0.2.0
jinja2
apscheduler
pyshorteners
colorama
fuzzywuzzy
python-Levenshtein==0.27.1

# Git-based Packages
https://github.com/DreamXBotz/cinemagoer/archive/master.zip
